refactor: change balance and amount to BigDecimal

+++ cash-machine/cash-machine-domain/src/main/java/com/example/cash/machine/domain/commands/DepositCommand.java
@@ -3,9 +3,11 @@ package com.example.cash.machine.domain.commands;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 public class DepositCommand {
 
-  private final int amount;
+  private final BigDecimal amount;
 }
